requirements.txt
Pillow==9.4.0
pyTelegramBotAPI==4.9.0
PyYAML==6.0
requests==2.28.2
schedule==1.1.0
vk_api==11.9.9
